+++ e2e/node/utils/mock-replica.ts
@@ -16,16 +16,26 @@ import {
   Signed,
   UnSigned,
   ReadStateRequest,
+  QueryRequest,
+  hashOfMap,
+  QueryResponseReplied,
+  IC_RESPONSE_DOMAIN_SEPARATOR,
+  QueryResponseStatus,
+  ReadRequestType,
 } from '@icp-sdk/core/agent';
 import { Principal } from '@icp-sdk/core/principal';
+import { Ed25519KeyIdentity } from '@icp-sdk/core/identity';
 import { Mock, vi } from 'vitest';
-import { createReplyTree, createTimeTree } from './tree';
+import { createReplyTree, createSubnetTree, createTimeTree } from './tree';
 import { randomKeyPair, signBls, KeyPair } from './identity';
-import { concatBytes } from '@noble/hashes/utils';
+import { concatBytes, toBytes } from '@noble/hashes/utils';
+
+const NANOSECONDS_TO_MSECS = 1_000_000;
 
 export enum MockReplicaSpyType {
   CallV3 = 'CallV3',
   ReadStateV2 = 'ReadStateV2',
+  QueryV2 = 'QueryV2',
 }
 
 export type MockReplicaRequest = Request<{ canisterId: string }, Uint8Array, Uint8Array>;
@@ -37,6 +47,15 @@ export type MockReplicaSpy = Mock<MockReplicaSpyImpl>;
 export interface MockReplicaSpies {
   [MockReplicaSpyType.CallV3]?: MockReplicaSpy;
   [MockReplicaSpyType.ReadStateV2]?: MockReplicaSpy;
+  [MockReplicaSpyType.QueryV2]?: MockReplicaSpy;
+}
+
+function fallbackSpyImpl(spyType: MockReplicaSpyType, canisterId: string): MockReplicaSpyImpl {
+  return (_req, res) => {
+    res
+      .status(500)
+      .send(`No implementation defined for ${spyType} spy on canisterId: ${canisterId}.`);
+  };
 }
 
 export class MockReplica {
@@ -54,6 +73,10 @@ export class MockReplica {
     app.post(
       '/api/v2/canister/:canisterId/read_state',
       this.#createEndpointSpy(MockReplicaSpyType.ReadStateV2),
+    );
+    app.post(
+      '/api/v2/canister/:canisterId/query',
+      this.#createEndpointSpy(MockReplicaSpyType.QueryV2),
     );
   }
 
@@ -84,12 +107,20 @@ export class MockReplica {
     this.#setSpyImplOnce(canisterId, MockReplicaSpyType.ReadStateV2, impl);
   }
 
+  public setV2QuerySpyImplOnce(canisterId: string, impl: MockReplicaSpyImpl): void {
+    this.#setSpyImplOnce(canisterId, MockReplicaSpyType.QueryV2, impl);
+  }
+
   public getV3CallSpy(canisterId: string): MockReplicaSpy {
     return this.#getSpy(canisterId, MockReplicaSpyType.CallV3);
   }
 
   public getV2ReadStateSpy(canisterId: string): MockReplicaSpy {
     return this.#getSpy(canisterId, MockReplicaSpyType.ReadStateV2);
+  }
+
+  public getV2QuerySpy(canisterId: string): MockReplicaSpy {
+    return this.#getSpy(canisterId, MockReplicaSpyType.QueryV2);
   }
 
   public getV3CallReq(canisterId: string, callNumber: number): Signed<CallRequest> {
@@ -102,6 +133,12 @@ export class MockReplica {
     const [req] = this.#getCallParams(canisterId, callNumber, MockReplicaSpyType.ReadStateV2);
 
     return Cbor.decode<UnSigned<ReadStateRequest>>(req.body);
+  }
+
+  public getV2QueryReq(canisterId: string, callNumber: number): UnSigned<QueryRequest> {
+    const [req] = this.#getCallParams(canisterId, callNumber, MockReplicaSpyType.QueryV2);
+
+    return Cbor.decode<UnSigned<QueryRequest>>(req.body);
   }
 
   #createEndpointSpy(spyType: MockReplicaSpyType): MockReplicaSpyImpl {
@@ -119,6 +156,9 @@ export class MockReplica {
         res.status(500).send(`No ${spyType} spy defined for canisterId: ${canisterId}.`);
         return;
       }
+
+      // add fallback implementation to return 500 if the spy runs out of implementations
+      spy.mockImplementation(fallbackSpyImpl(spyType, canisterId));
 
       spy(req, res);
     };
@@ -258,7 +298,7 @@ export async function prepareV3Response({
   };
 
   return {
-    responseBody: new Uint8Array(Cbor.encode(responseBody)),
+    responseBody: Cbor.encode(responseBody),
     requestId,
   };
 }
@@ -298,8 +338,165 @@ export async function prepareV2ReadStateTimeResponse({
   };
 
   return {
-    responseBody: new Uint8Array(Cbor.encode(responseBody)),
+    responseBody: Cbor.encode(responseBody),
   };
+}
+
+interface V2ReadStateSubnetOptions {
+  nodeIdentity: Ed25519KeyIdentity;
+  canisterRanges: Array<[Uint8Array, Uint8Array]>;
+  keyPair?: KeyPair;
+  date?: Date;
+}
+
+/**
+ * Prepares a version 2 read state subnet response.
+ * @param {V2ReadStateSubnetOptions} options - The options for preparing the response.
+ * @param {Ed25519KeyIdentity} options.nodeIdentity - The identity of the node.
+ * @param {Array<[Uint8Array, Uint8Array]>} options.canisterRanges - The canister ranges for the subnet.
+ * @param {KeyPair} options.keyPair - The key pair for signing.
+ * @param {Date} options.date - The date for the response.
+ * @returns {Promise<V2ReadStateResponse>} A promise that resolves to the prepared response.
+ */
+export async function prepareV2ReadStateSubnetResponse({
+  nodeIdentity,
+  canisterRanges,
+  keyPair,
+  date,
+}: V2ReadStateSubnetOptions): Promise<V2ReadStateResponse> {
+  keyPair = keyPair ?? randomKeyPair();
+  date = date ?? new Date();
+
+  const subnetId = Principal.selfAuthenticating(keyPair.publicKeyDer).toUint8Array();
+
+  const tree = createSubnetTree({
+    subnetId,
+    nodeIdentity,
+    canisterRanges,
+    date,
+  });
+  const signature = await signTree(tree, keyPair);
+
+  const cert: Cert = {
+    tree,
+    signature,
+  };
+  const responseBody: ReadStateResponse = {
+    certificate: Cbor.encode(cert),
+  };
+
+  return {
+    responseBody: Cbor.encode(responseBody),
+  };
+}
+
+interface V2QueryResponseOptions {
+  canisterId: Principal | string;
+  methodName: string;
+  arg: Uint8Array;
+  sender: Principal | string;
+  nodeIdentity: Ed25519KeyIdentity;
+  ingressExpiryInMinutes?: number;
+  timeDiffMsecs?: number;
+  reply?: string | Uint8Array;
+  date?: Date;
+}
+
+interface V2QueryResponse {
+  responseBody: Uint8Array;
+  requestId: RequestId;
+}
+
+/**
+ * Prepares a version 2 query response.
+ * @param {V2QueryResponseOptions} options - The options for preparing the response.
+ * @param {string} options.canisterId - The ID of the canister.
+ * @param {string} options.methodName - The name of the method being called.
+ * @param {Uint8Array} options.arg - The arguments for the method call.
+ * @param {string} options.sender - The principal ID of the sender.
+ * @param {Ed25519KeyIdentity} options.nodeIdentity - The identity of the node.
+ * @param {number} options.ingressExpiryInMinutes - The ingress expiry time in minutes.
+ * @param {number} options.timeDiffMsecs - The time difference in milliseconds.
+ * @param {Uint8Array} options.reply - The reply payload.
+ * @param {Date} options.date - The date for the response.
+ * @returns {Promise<V2QueryResponse>} A promise that resolves to the prepared response.
+ */
+export async function prepareV2QueryResponse({
+  canisterId,
+  methodName,
+  arg,
+  sender,
+  nodeIdentity,
+  ingressExpiryInMinutes,
+  timeDiffMsecs,
+  reply,
+  date,
+}: V2QueryResponseOptions): Promise<V2QueryResponse> {
+  canisterId = Principal.from(canisterId);
+  sender = Principal.from(sender);
+  ingressExpiryInMinutes = ingressExpiryInMinutes ?? 5;
+  timeDiffMsecs = timeDiffMsecs ?? 0;
+  const coercedReply = reply ? toBytes(reply) : new Uint8Array();
+  date = date ?? new Date();
+
+  const ingressExpiry = calculateIngressExpiry(ingressExpiryInMinutes, timeDiffMsecs);
+  const queryRequest: QueryRequest = {
+    request_type: ReadRequestType.Query,
+    canister_id: canisterId,
+    method_name: methodName,
+    arg,
+    sender,
+    ingress_expiry: ingressExpiry,
+  };
+
+  const requestId = requestIdOf(queryRequest);
+  const timestamp = BigInt(date.getTime()) * BigInt(NANOSECONDS_TO_MSECS);
+
+  const message = createQueryReplyMessage({
+    requestId,
+    status: QueryResponseStatus.Replied,
+    reply: coercedReply,
+    timestamp,
+  });
+  const signature = await nodeIdentity.sign(message);
+
+  const body: QueryResponseReplied = {
+    status: QueryResponseStatus.Replied,
+    reply: { arg: coercedReply },
+    signatures: [
+      {
+        timestamp,
+        signature,
+        identity: nodeIdentity.getPrincipal().toUint8Array(),
+      },
+    ],
+  };
+
+  return {
+    responseBody: Cbor.encode(body),
+    requestId,
+  };
+}
+
+function createQueryReplyMessage({
+  requestId,
+  status,
+  reply,
+  timestamp,
+}: {
+  requestId: RequestId;
+  status: QueryResponseStatus;
+  reply: Uint8Array;
+  timestamp: bigint;
+}): Uint8Array {
+  const hash = hashOfMap({
+    status,
+    reply: { arg: reply },
+    timestamp,
+    request_id: requestId,
+  });
+
+  return concatBytes(IC_RESPONSE_DOMAIN_SEPARATOR, hash);
 }
 
 async function signTree(tree: HashTree, keyPair: KeyPair): Promise<Uint8Array> {